refactor: rename queryApi to askApi and move Fisherman tools into src/ai/fisherman/
queryApi read as if the tool ran a live query against the API. It does not:
in replicate mode Fisherman works from captured traffic, and otherwise it
guesses from the spec definition. askApi says what actually happens — a
question is asked, and the answer may not come.

fisherman-tools.ts moves to src/ai/fisherman/tools.ts alongside request-haul,
and the tool definition moves out of pilot.ts into it. The file now separates
the two audiences by export: createFishermanTools for the tools Fisherman
calls inside its own session, createAskApiTool for the one other agents offer.
