+++ CHANGELOG.md
@@ -7,7 +7,7 @@
 - [Fisherman] Can now answer questions about data that already exists, over read-only GET requests
   that create or change nothing. It draws on the same endpoints it already knew about — an OpenAPI
   spec when one is configured, or endpoints learned by watching browser traffic otherwise.
-- [Pilot] Gained a `queryApi(question)` tool alongside `precondition()`, available while planning a
+- [Pilot] Gained an `askApi(question)` tool alongside `precondition()`, available while planning a
   test, reviewing a new page, and checking progress mid-run. It asks Fisherman what already exists —
   whether suitable data is already there, or the exact name or id of an existing record — before
   deciding whether to create anything.
